+++ test/fixtures/conf/expected/scripts.js
@@ -1,1 +1,0 @@
-var name="World",message="Hello ,"+name+"!";console.log(message);